Squashed commit of the following:
commit 08b43bd481019f5cddc73d1ad8c8261b56be1e2c
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 12:58:31 2022 +0100

    Release work.

commit 20901f1e9ab8f0ebe82a2f07526bf344890736f8
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 12:56:36 2022 +0100

    Release work

commit 697eeb7f680fdf90e045c4aef406f65a0fa75e24
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 12:54:30 2022 +0100

    Release work.

commit 81967b979c649c6311df961c194fe3f6efc1ba49
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 12:13:15 2022 +0100

    Package generation for linux.

commit 30f96bd9a643a0d26f43e5ec6ab97ba1e0cc5952
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 11:49:22 2022 +0100

    CI Work

commit 62cd944693e3205b265f39913c9a82d5c7762753
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 11:48:21 2022 +0100

    CI Work

commit 9ffc198ca6deab680cd07ef7868e861c8644522d
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 11:47:39 2022 +0100

    Add linux to CI.

commit 98d83a8f43ac67ff4af0fc9605600276bb2df20c
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sun Apr 17 11:44:41 2022 +0100

    Now running correctly on linux.

commit 8e3c8642f9e4624338919efe686d8d9a3bea6e22
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sat Apr 16 22:29:25 2022 +0100

    Linking on linux. Just need to fix up some errors with copying dependent DLL's to output directory.

commit 71c1de25db2f2a9f1434baa14deebd3f9712dd04
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sat Apr 16 22:08:12 2022 +0100

    Keep windows build happy.

commit dfe3ecc6227d2cdce8f54671588b38a5eb02d8c9
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sat Apr 16 22:05:37 2022 +0100

    Compiles for linux, couple of linker errors to fix.

commit 67695e49644c419ba1760cd016b9ead17a47d794
Author: Tim Leonard <tim@programmingbytim.com>
Date:   Sat Apr 16 17:27:58 2022 +0100

    Moved over some initial linux fixes from @garyttierney

+++ Source/Server/Server/GameService/GameManagers/Signs/SignManager.cpp
@@ -24,6 +24,8 @@
 #include "Core/Utils/Strings.h"
 #include "Core/Utils/DiffTracker.h"
 
+#include <cmath>
+
 SignManager::SignManager(Server* InServerInstance, GameService* InGameServiceInstance)
     : ServerInstance(InServerInstance)
     , GameServiceInstance(InGameServiceInstance)
@@ -533,7 +535,7 @@ MessageHandleResult SignManager::Handle_RequestGetRightMatchingArea(GameClient* 
     // Normalize the values to the 0-5 range the client expects and return them.
     for (auto Pair : PotentialAreas)
     {
-        int NormalizedPopulation = (int)std::ceilf((Pair.second / (float)MaxAreaPopulation) * 5.0f);
+        int NormalizedPopulation = (int)std::ceil((Pair.second / (float)MaxAreaPopulation) * 5.0f);
 
         Frpg2RequestMessage::RequestGetRightMatchingAreaResponse_Area_info& Info = *Response.add_area_info();
         Info.set_online_area_id((uint32_t)Pair.first);
